Input String Fix 5: Final
Found a much simpler way to handle this fix, added method to main game class and use it to convert input before sending it to all other methods. Had to go back and revert all the changes I've made this week but this is a much simpler solution. Final commit fixes all known bugs.